[Bench][Pool] Time every row against the strongest baseline, and fold the file in
torch through inductor is the strongest pooling baseline for nine of the twelve ops, by
2-6x over eager on 1D, and nothing here was measuring it. Every row now times four
implementations — tileops, the library where one applies, eager torch, and compiled torch.
The strongest is per row, not per op: inductor is 2.7x worse than eager on one 3D max-pool
shape and 6x worse on another, so both torch readings stay.

dynamo caches eight graphs per code object and all 48 cases share one ref_program, so the
compiled baseline resets dynamo per case. Without that the later cases would time eager
under a tag that says compiled.

_pool_baselines.py is gone; its contents live in bench_pool.py. benchmarks/ops holds one
file per op family and no other helper module, and the adapters are only reachable from
this family's cases.
